renamed main to __init__
